requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
twilio==9.7.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.3
